backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
pydantic
python-dotenv
aiohttp
requests
pytest
pytest-aysncio
httpx
asgi-lifespan